Refactored game-landing and game canvas into seperate module.
