+++ src/app/app.module.ts
@@ -2,36 +2,33 @@ import './rxjs-extensions';
 
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
-import { HttpModule } from '@angular/http';
 
 import { AppRoutingModule } from './app-routing.module';
+import { CoreModule } from './core/core.module';
+import { GameLandingModule } from './game-landing/game-landing.module';
 
 import { AppComponent } from './app.component';
-import { GameLandingComponent } from './game-landing/game-landing.component';
 import { SiteLandingComponent } from './site-landing/site-landing.component';
 import { ContactInfoComponent } from './contact-info/contact-info.component';
 import { TechUsedComponent } from './tech-used/tech-used.component';
-import { GameCanvasComponent } from './game-canvas/game-canvas.component';
-import { ScoreService }  from './score.service';
 
 @NgModule({
 
   imports: [
+    AppRoutingModule,
     BrowserModule,
-    FormsModule,
-    HttpModule,
-    AppRoutingModule
+    CoreModule,
+    GameLandingModule
+
+
   ],
   declarations: [
     AppComponent,
-    GameLandingComponent,
     SiteLandingComponent,
     ContactInfoComponent,
-    TechUsedComponent,
-    GameCanvasComponent,
+    TechUsedComponent
   ],
-  providers: [ScoreService],
-  bootstrap: [AppComponent]
+  providers: [],
+  bootstrap: [ AppComponent ]
 })
 export class AppModule { }
