+++ src/app/game-landing/score.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Headers, Http } from '@angular/http';
 import { Score } from './Score';
+
 import 'rxjs/add/operator/toPromise';
 
 @Injectable()
